Rework API to have coordination work in a single place
